+++ components/FolderList/FolderList.tsx
@@ -1,8 +1,8 @@
 import React, { useState, useEffect } from "react";
-import { useFetchFolders } from "../hooks/useFetchFolders";
-import { fetchFolders } from "../api/fetchFolders";
+import { useFetchFolders } from "../../hooks/useFetchFolders";
+import { fetchFolders } from "../../api/fetchFolders";
 import FolderContent from "../FolderContent/FolderContent";
-import { Link } from "../interfaces/model";
+import { Link } from "../../interfaces/model";
 
 interface FolderListProps {
   searchTerm: string;
