+++ hooks/ useFetchCardsData.tsx
@@ -1,6 +1,6 @@
 import { useState, useEffect } from "react";
-import { calculateTimePassed, formatDate } from "../ utils";
-import { CardDataProps } from "../Cardlist/Cardlist";
+import { calculateTimePassed, formatDate } from "../components/ utils";
+import { CardDataProps } from "../components/Cardlist/Cardlist";
 
 export function useFetchCardsData() {
     const [cardsData, setCardsData] = useState<CardDataProps[]>([]);
