+++ interfaces/model.ts
